dafitishoes/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
gunicorn==20.0.4
ipython==7.21.0
djangorestframework
redis==3.4.1
celery==4.4.1